マイグレーションSQLのgenerate_series呼び出しに型キャストを付け、EXPLAIN取得を可能にする
pg_stat_statementsがリテラル引数も$Nパラメータに正規化するため、
nrpostgresqlreceiver(top_query_collection)がEXPLAIN実行時にgenerate_seriesの
オーバーロード解決に失敗し"function generate_series(unknown, unknown) is not
unique"で失敗していた。

+++ migrations/20240103000000_add_company_id_to_workflow_definitions.sql
@@ -29,22 +29,22 @@ ON workflow_definitions(company_id);
 -- 6. 各会社（1-50）ごとにワークフロー定義を作成
 -- 出張申請（BusinessTrip）
 INSERT INTO workflow_definitions (application_type, company_id)
-SELECT 'BusinessTrip', generate_series(1, 50)
+SELECT 'BusinessTrip', generate_series(1::integer, 50::integer)
 ON CONFLICT (application_type, company_id) DO NOTHING;
 
 -- 経費申請（Expense）
 INSERT INTO workflow_definitions (application_type, company_id)
-SELECT 'Expense', generate_series(1, 50)
+SELECT 'Expense', generate_series(1::integer, 50::integer)
 ON CONFLICT (application_type, company_id) DO NOTHING;
 
 -- 休暇申請（Vacation）
 INSERT INTO workflow_definitions (application_type, company_id)
-SELECT 'Vacation', generate_series(1, 50)
+SELECT 'Vacation', generate_series(1::integer, 50::integer)
 ON CONFLICT (application_type, company_id) DO NOTHING;
 
 -- プロモーション申請（Promotion）
 INSERT INTO workflow_definitions (application_type, company_id)
-SELECT 'Promotion', generate_series(1, 50)
+SELECT 'Promotion', generate_series(1::integer, 50::integer)
 ON CONFLICT (application_type, company_id) DO NOTHING;
 
 -- 7. 各会社ごとにワークフローステップを作成
